+++ apps/app/src/main.ts
@@ -37,7 +37,8 @@ const { provider, onReady } = createProvider({
   },
   project: {
     id: name
-  }
+  },
+  enableStaticRoute: true
 });
 
 onReady(async () => {