+++ packages/parser/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.67](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.66...@vtj/parser@0.12.67) (2025-07-31)
+
+
+### Bug Fixes
+
+* 🐛 parser TEXT_CALL ([f3fa726](https://gitee.com/newgateway/vtj/commits/f3fa726ce59a2a8b96da93e4415dd4b50dfb9bca))
+
+
+
+
+
 ## [0.12.66](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.65...@vtj/parser@0.12.66) (2025-07-26)